improve handling of URLs and also remove FORWARD rules from iptables scripts
The FORWARD rules don't seem to be needed.

+++ src/armour-master/src/lib.rs
@@ -10,39 +10,48 @@ fn string_from_bytes(b: bytes::Bytes) -> String {
 }
 
 pub async fn control_plane<S: serde::Serialize>(
-    url: &str,
+    url: &url::Url,
     method: http::Method,
     path: &str,
     value: &S,
 ) -> Result<String, String> {
-    let client = actix_web::client::Client::default();
-    let full_path = format!("http://{}/{}", url, path);
-    let req = match method {
-        http::Method::GET => client.get(full_path),
-        http::Method::POST => client.post(full_path),
-        http::Method::DELETE => client.delete(full_path),
-        _ => return Err("bad method".into()),
-    };
-    match req.send_json(value).await {
-        Ok(mut res) => {
-            let body = string_from_bytes(
-                res.body()
-                    .await
-                    .map_err(|_| "failed to read body".to_string())?,
-            );
-            // log::debug!("{:?}", body);
-            if res.status().is_success() {
-                Ok(body)
-            } else {
-                Err(body)
+    if let Some(host_str) = url.host_str() {
+        let client = actix_web::client::Client::default();
+        let full_path = format!(
+            "http://{}:{}/{}",
+            host_str,
+            url.port().unwrap_or(8088),
+            path
+        );
+        let req = match method {
+            http::Method::GET => client.get(full_path),
+            http::Method::POST => client.post(full_path),
+            http::Method::DELETE => client.delete(full_path),
+            _ => return Err("bad method".into()),
+        };
+        match req.send_json(value).await {
+            Ok(mut res) => {
+                let body = string_from_bytes(
+                    res.body()
+                        .await
+                        .map_err(|_| "failed to read body".to_string())?,
+                );
+                // log::debug!("{:?}", body);
+                if res.status().is_success() {
+                    Ok(body)
+                } else {
+                    Err(body)
+                }
             }
+            Err(err) => Err(format!("{}: {}", url, err)),
         }
-        Err(err) => Err(format!("{}: {}", url, err)),
+    } else {
+        Err(format!("bad control plane URL: {}", url))
     }
 }
 
 pub async fn control_plane_deserialize<S, D>(
-    url: &str,
+    url: &url::Url,
     method: http::Method,
     path: &str,
     value: &S,
